Analyze the GetProblem
- The model of concurrency used in the proposed example was really classically obsolete regarding Java 8 features
- The new model of testing such bug:
	- Spawn a huge number of messages to the actor at once
	- Collect all the future values
	- Count the number of all successful futures completed
	- Verify that it is the same as input number

Tested with numbers around 5M. To test, make sure you have Java 8 on the path.
Simpley run `test.sh`.
