+++ .gitignore
@@ -1,0 +1,6 @@
+*.class
+target/*
+.project
+.classpath
+.settings/*
+
